backend/requirements.txt
# RAG Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# RAG Core
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langchain-openai>=0.3.0
langchain-deepseek>=0.1.0
langchain-ollama>=0.0.4
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
pypdf>=3.17.0
python-docx>=1.1.0

# Optional (按需安装)
# langchain-milvus>=0.1.0
# langchain-cohere>=0.1.0
